refactor(mention): move utils function to another file
